requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
google-generativeai==0.4.1